requirements.txt
pystan==2.19.1.1
fbprophet
pandas
numpy
yfinance
plotly
matplotlib 
streamlit